+++ Sources/ToucanSDK/ContentType/ContentType.swift
