+++ Sources/ToucanSDK/Context/RedirectContext.swift
@@ -1,0 +1,11 @@
+//
+//  File.swift
+//  toucan
+//
+//  Created by Tibor Bodecs on 2024. 10. 13..
+//
+
+struct RedirectContext {
+    let url: String
+    let code: Int
+}
