+++ Sources/ToucanSDK/Context/SitemapContext.swift
@@ -5,7 +5,7 @@
 //  Created by Tibor Bodecs on 14/05/2024.
 //
 
-struct Sitemap: Output {
+struct SitemapContext {
 
     struct URL {
         let location: String
