--- Sources/ToucanSDK/Site/Types/Output+Redirect.swift
@@ -1,11 +1,0 @@
-//
-//  Context+Redirect.swift
-//
-//
-//  Created by Tibor Bodecs on 01/07/2024.
-//
-
-struct Redirect: Output {
-    let url: String
-    let code: Int
-}
