--- Sources/ToucanSDK/Site/Types/Output.swift
@@ -1,8 +1,0 @@
-//
-//  File.swift
-//
-//
-//  Created by Tibor Bodecs on 20/06/2024.
-//
-
-protocol Output {}
